Move planner to federation utils
